pyproject.toml
[tool.poetry]
name = "realtime-fraud-detection"
version = "0.1.0"
description = "Production-grade real-time fraud detection using PySpark, Kinesis, and SageMaker"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.9"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.34.34"
faker = "^22.6.0"
pandas = "^2.2.0"
numpy = "^1.26.3"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.1"
mypy = "^1.8.0"
flake8 = "^7.0.0"
isort = "^5.13.2"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
